move EditWhisky into its own file
